server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">= 3.10"
dependencies = [
    "langchain-community>=0.3.1",
    "langgraph>=0.2.32",
    "starlette>=0.39.2",
    "uvicorn[standard]>=0.31.0",
    # Additional runtime dependencies required by server tools
    "geocoder>=1.38.1",
    "httpx>=0.24.1",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py310"
